// 25-Ecommerce-Project/code-examples/angular-ecommerce/src/app/core/models/user.model.ts
export interface User {
  id: string;
  email: string;
  name: string;
  isAdmin: boolean;
  avatar?: string;
  phone?: string;
  addresses?: Address[];
  defaultAddressId?: string;
  createdAt: string;
  updatedAt: string;
}

export interface Address {
  id: string;
  fullName: string;
  addressLine1: string;
  addressLine2?: string;
  city: string;
  state: string;
  postalCode: string;
  country: string;
  phone: string;
  isDefault?: boolean;
}

export interface AuthResponse {
  user: User;
  token: string;
  expiresIn: number;
}